光の円柱を描画。地形を貫通してエミッシブが見えるバグを修正。

+++ src/engine/Render/RenderObject/ModelInfo/ModelAnimator.cpp
@@ -243,7 +243,7 @@ void KuroEngine::ModelAnimator::Play(const std::string& AnimationName, const boo
 	//Ä¶’†‚È‚çƒŠƒZƒbƒg‚µ‚Ä‚¨‚í‚è
 	if (already != playAnimations.end())
 	{
-		already->past = InitPast;
+		already->past = static_cast<float>(InitPast);
 		already->loop = Loop;
 		return;
 	}
